+++ src/Moryx.Products.Management/Implementations/Storage/ProductStorage.cs
@@ -7,6 +7,7 @@ using Moryx.Container;
 using Moryx.Model;
 using Moryx.Model.Repositories;
 using Moryx.Products.Management.Components;
+using Moryx.Products.Management.ModuleControllers;
 using Moryx.Products.Model;
 using Moryx.Tools;
 using System;
