+++ src/Moryx.Products.Management/Modifications/Models/ImportStateModel.cs
@@ -4,7 +4,7 @@
 using Moryx.Products.Management.Implementations;
 using System.Runtime.Serialization;
 
-namespace Moryx.Products.Management.Modification
+namespace Moryx.Products.Management.Modification.Models
 {
     [DataContract]
     internal class ImportStateModel
