+++ src/Moryx.Products.Management/Modifications/Models/PartConnector.cs
@@ -4,10 +4,10 @@
 using Moryx.Serialization;
 using System.Runtime.Serialization;
 
-namespace Moryx.Products.Management.Modification
+namespace Moryx.Products.Management.Modification.Models
 {
     [DataContract]
-    internal class ProductDefinitionModel
+    internal class PartConnector
     {
         [DataMember]
         public string Name { get; set; }
@@ -16,9 +16,15 @@ namespace Moryx.Products.Management.Modification
         public string DisplayName { get; set; }
 
         [DataMember]
-        public string BaseDefinition { get; set; }
+        public string Type { get; set; }
 
         [DataMember]
-        public Entry Properties { get; set; }
+        public bool IsCollection { get; set; }
+
+        [DataMember]
+        public PartModel[] Parts { get; set; }
+
+        [DataMember]
+        public Entry PropertyTemplates { get; set; }
     }
 }
