+++ src/Moryx.Products.Management/Modifications/Models/PartModel.cs
@@ -4,7 +4,7 @@
 using Moryx.Serialization;
 using System.Runtime.Serialization;
 
-namespace Moryx.Products.Management.Modification
+namespace Moryx.Products.Management.Modification.Models
 {
     [DataContract]
     internal class PartModel
