+++ src/Moryx.Products.Management/Modifications/Models/ProductImporter.cs
@@ -1,17 +1,18 @@
 // Copyright (c) 2020, Phoenix Contact GmbH & Co. KG
 // Licensed under the Apache License, Version 2.0
 
+using Moryx.Serialization;
 using System.Runtime.Serialization;
 
-namespace Moryx.Products.Management.Modification
+namespace Moryx.Products.Management.Modification.Models
 {
     [DataContract]
-    internal class StorageValue
+    internal class ProductImporter
     {
         [DataMember]
-        public string Key { get; set; }
+        public string Name { get; set; }
 
         [DataMember]
-        public string[] Values { get; set; }
+        public Entry Parameters { get; set; }
     }
 }
