+++ src/Moryx.Products.Management/Modifications/Models/ProductModel.cs
@@ -7,7 +7,7 @@ using Moryx.AbstractionLayer.Products;
 using Moryx.Serialization;
 using System.Runtime.Serialization;
 
-namespace Moryx.Products.Management.Modification
+namespace Moryx.Products.Management.Modification.Models
 {
     [DataContract]
     internal class ProductModel
