+++ src/Moryx.Products.Management/Modifications/Models/RecipeClassificationModel.cs
@@ -1,7 +1,7 @@
 // Copyright (c) 2020, Phoenix Contact GmbH & Co. KG
 // Licensed under the Apache License, Version 2.0
 
-namespace Moryx.Products.Management.Modification
+namespace Moryx.Products.Management.Modification.Models
 {
     /// <summary>
     /// Webservice compatible recipe classification enum
