+++ src/Moryx.Products.Management/Modifications/Models/StorageValue.cs
@@ -1,18 +1,17 @@
 // Copyright (c) 2020, Phoenix Contact GmbH & Co. KG
 // Licensed under the Apache License, Version 2.0
 
-using Moryx.Serialization;
 using System.Runtime.Serialization;
 
-namespace Moryx.Products.Management.Modification
+namespace Moryx.Products.Management.Modification.Models
 {
     [DataContract]
-    internal class ProductImporter
+    internal class StorageValue
     {
         [DataMember]
-        public string Name { get; set; }
+        public string Key { get; set; }
 
         [DataMember]
-        public Entry Parameters { get; set; }
+        public string[] Values { get; set; }
     }
 }
