+++ src/Moryx.Products.Management/Modifications/Models/WorkplanModel.cs
@@ -26,7 +26,7 @@ using System.Runtime.Workflows;
 using Moryx.Workflows;
 using System.Runtime.Serialization;
 
-namespace Moryx.Products.Management.Modification
+namespace Moryx.Products.Management.Modification.Models
 {
     [DataContract]
     internal class WorkplanModel
