+++ src/Moryx.Products.Management/Plugins/GenericStrategies/GenericTypeStrategy.cs
@@ -8,6 +8,7 @@ using Moryx.Products.Management.Components;
 using Moryx.Products.Management.Implementations.Storage;
 using Moryx.Products.Model;
 using System;
+using System.Linq.Expressions;
 /* Unmerged change from project 'Moryx.Products.Management (net5.0)'
 Before:
 using System;
