+++ src/Moryx.Products.Management/Plugins/GenericStrategies/IntegerColumnMapper.cs
@@ -28,6 +28,8 @@ using Moryx.Tools;
 using System;
 */
 using Moryx.Container;
+using Moryx.Products.Management.Implementations.Storage;
+using Moryx.Tools;
 
 /* Unmerged change from project 'Moryx.Products.Management (net5.0)'
 Before:
@@ -51,6 +53,7 @@ using System.Reflection;
 using System.Reflection;
 */
 using System;
+using System.Reflection;
 
 namespace Moryx.Products.Management.Plugins.GenericStrategies
 {
