+++ src/Moryx.Products.Samples/Importer/WatchImporter.cs
@@ -29,6 +29,7 @@ using Moryx.Products.Management.Components;
 using Moryx.Products.Management.Implementations.Import;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
+using System.Threading.Tasks;
 /* Unmerged change from project 'Moryx.Products.Samples (netcoreapp3.1)'
 Before:
 using System.Collections.Generic;
