+++ src/Moryx.Products.Samples/Strategies/WatchStrategy.cs
@@ -7,6 +7,7 @@ using Moryx.Products.Management.Components;
 using Moryx.Products.Management.Implementations.Storage;
 using Moryx.Products.Model;
 using System;
+using System.Linq.Expressions;
 /* Unmerged change from project 'Moryx.Products.Samples (netcoreapp3.1)'
 Before:
 using System;
