+++ src/Moryx.Resources.Management/Resources/IResourceLinker.cs
@@ -5,6 +5,7 @@ using Moryx.AbstractionLayer.Resources;
 using Moryx.Model.Repositories;
 using Moryx.Resources.Model.Entities;
 using System.Collections.Generic;
+using System.Reflection;
 /* Unmerged change from project 'Moryx.Resources.Management (net5.0)'
 Before:
 using System.Collections.Generic;
