+++ src/Moryx.Resources.Management/Resources/ResourceLinker.cs
@@ -39,6 +39,9 @@ using Moryx.Tools;
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.Linq;
+using System.Reflection;
 /* Unmerged change from project 'Moryx.Resources.Management (net5.0)'
 Before:
 using System;
